+++ src/main/java/learning/java/game/dao/GameDao.java
@@ -1,6 +1,6 @@
-package learning.java.game.DAO;
+package learning.java.game.dao;
 
-import learning.java.game.gameModel.model.Game;
+import learning.java.game.game_model.model.Game;
 import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
@@ -8,12 +8,13 @@ import java.util.Map;
 import java.util.UUID;
 
 @Component
-public class DAO {
+public class GameDao {
     private Map<UUID, Game> dataBase = new HashMap<>();
 
     public Map<UUID, Game> getDataBase() {
         return dataBase;
     }
+
     public void setDataBase(UUID id, Game game) {
         dataBase.put(id, game);
     }
